// sdk/src/pins.ts
// Generated by agoraseal sdk-pin; normal gates only --check.
export const pins = {
  sp1: '0xa14f73480477fb0d2bddec5b9d7df38ad03c2ca861441d66de00862043e85b02',
  stateContext: '0xbab53b4c8d28033a9194b8b5acd08a062ae712ae03bdd1d1d8196106b8ccfe0d',
  voteContext: '0x5739d9d4928ee4ed0abe51456f708fb155e7d59cfe3d76f06980b1b3060255c8',
  vote: '0xe5090fbbe406884866f57b51a22e41e6087fdb5195d09fc9b957e188f8a9bb74',
  proposal: '0x1c5cfaeaee7d318c13d5882d079fb39a58ba503ebb20efc355da346c2c0568cd',
  treasury: '0x426fc417f5402a7d5ca78363357b9de31fbd9496806ad59249d2c5f3fc3a78d9',
} as const;
export const programKey = '0x00e6a25086d9a6ba917d12d14e857f93f57b1cbfd828466c2e352c419f848502';